+++ lib/github/model/tag.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Kraken
+  module GitHub
+    class Tag
+      attr_accessor :name
+      def initialize(name)
+        @name = name
+      end
+    end
+  end
+end