+++ spec/github/wrapper_spec.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-require 'github/client_wrapper'
+require 'github/wrapper'
 
-RSpec.describe GitHub::ClientWrapper do
+RSpec.describe Kraken::GitHub::Wrapper do
   let(:client) { double }
 
-  subject { GitHub::ClientWrapper.new(client) }
+  subject { Kraken::GitHub::Wrapper.new(client) }
 
   it 'delegates list_tags to the initialized client' do
     expect(client).to receive(:refs).with('burrito/taco-server')
@@ -16,15 +16,15 @@ RSpec.describe GitHub::ClientWrapper do
                                                 ])
 
     tags = subject.list_tags('burrito', 'taco-server')
-    expect(tags).to eq(%w[v0.1.0 v0.2.0])
+    expect(tags.map(&:name)).to eq(%w[v0.1.0 v0.2.0])
   end
 
   private
 
   def double_ref(tag)
-    ref = double
-    expect(ref).to receive(:ref).at_least(1).times
-                                .and_return(tag)
-    ref
+    fake_ref = double
+    expect(fake_ref).to receive(:ref).at_least(1).times
+                                     .and_return(tag)
+    fake_ref
   end
 end