+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Nothing
 
+## [0.6.0] - March 22, 2020
+
+### Added
+- Model for Trello cards, GitHub tags and Kubernetes pods
+
+### Updated
+- Cleaned up the module approach for all client wrappers
+
 ## [0.5.0] - March 22, 2020
 
 ### Added
@@ -40,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial project creation
 - Added rubocop
 
-[Unreleased]: https://github.com/jmtrusona/kraken/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/jmtrusona/kraken/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/jmtrusona/kraken/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/jmtrusona/kraken/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/jmtrusona/kraken/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/jmtrusona/kraken/compare/v0.2.0...v0.3.0